+++ cli/cmd/install.go
@@ -1,4 +1,4 @@
-// Copyright 2019 NetApp, Inc. All Rights Reserved.
+// Copyright 2020 NetApp, Inc. All Rights Reserved.
 package cmd
 
 import (
@@ -20,6 +20,7 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	policy "k8s.io/api/policy/v1beta1"
+	apiextensionv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	"k8s.io/client-go/rest"
 
 	"github.com/netapp/trident/cli/api"
@@ -500,6 +501,7 @@ func installTrident() (returnError error) {
 		pvcExists     bool
 		pvExists      bool
 		migrateToCRDs bool
+		crd           *apiextensionv1beta1.CustomResourceDefinition
 	)
 
 	// Ensure legacy Trident isn't already installed
@@ -535,6 +537,44 @@ func installTrident() (returnError error) {
 		log.WithField("namespace", TridentPodNamespace).Debug("Namespace does not exist.")
 	}
 
+	// Check for alpha snapshot CRDs
+	crdNames := []string{
+		"volumesnapshotclasses.snapshot.storage.k8s.io",
+		"volumesnapshotcontents.snapshot.storage.k8s.io",
+		"volumesnapshots.snapshot.storage.k8s.io",
+	}
+
+	for _, crdName := range crdNames {
+		// See if CRD exists
+		crdsExist, returnError = client.CheckCRDExists(crdName)
+		if returnError != nil {
+			return
+		}
+		if !crdsExist {
+			log.WithField("CRD", crdName).Debug("CRD not present.")
+			continue
+		}
+
+		// Get the CRD and check version
+		crd, returnError = client.GetCRD(crdName)
+		if returnError != nil {
+			return
+		}
+		alphaSnapshotErrorString := "kubernetes snapshot beta feature is not backwards compatible; run `tridentctl" +
+			" obliviate alpha-snapshot-crd` to remove previous kubernetes snapshot CRDs, " +
+			"then retry installation; for details, please refer to Trident’s online documentation"
+		if strings.ToLower(crd.Spec.Version) == "v1alpha1" {
+			returnError = fmt.Errorf(alphaSnapshotErrorString)
+			return
+		}
+		for _, version := range crd.Spec.Versions {
+			if strings.ToLower(version.Name) == "v1alpha1" {
+				returnError = fmt.Errorf(alphaSnapshotErrorString)
+				return
+			}
+		}
+	}
+
 	// Discover CRD data
 	crdsExist, returnError = client.CheckCRDExists(VersionCRDName)
 	if returnError != nil {
@@ -542,7 +582,7 @@ func installTrident() (returnError error) {
 	}
 
 	if crdsExist {
-		log.Debug("CRDs present, skipping PVC/PV check.")
+		log.Debug("Trident CRDs present, skipping PVC/PV check.")
 	} else {
 
 		// We didn't find any CRD data, so look for legacy etcd data